finally got the rendering to make sense
Turns out the space passes in the compile were over-applying the
model->clipspace matrix. This mangled the results. HOWEVER hunting for
that did help me find a tonne of bugs in the routing code so
c'est la vie

Got a lot of crap to clean up now. I also want to make the component
system less anal about access to component data.