haha what a mess!
Ok so we can see some shit, it is totally corrupt but we can work on
that. There is still this weird ass stuff with fbos though. We need to
look into unbinding fbos. I though we did this properly though..one sec,
I'll go check. Ok so fbo unbinds at the end..and im 100% sure
that attachments do too..Ah ok it's blending that is unbound. So we do
still need to look into %gl:draw-buffers and by extension
%fbo-draw-buffers.

It's getting exciting though. Even if we have two machines with rather
different output :|